+++ Sandswept/Enemies/SwepSwepTheSandy/SwepSwepTheSandy.cs
@@ -1,9 +1,9 @@
 using EntityStates.QuestVolatileBattery;
 using static RoR2.GivePickupsOnStart;
 
-namespace Sandswept.Enemies.SwepSwep
+namespace Sandswept.Enemies.SwepSwepTheSandy
 {
-    public class SwepSwep
+    public class SwepSwepTheSandy
     {
         public static void Init()
         {
@@ -11,12 +11,49 @@ namespace Sandswept.Enemies.SwepSwep
             var anyEquipmentDisabled = Main.AllEquipment.Count < Main.EnabledEquipment.Count;
             if (anyItemDisabled || anyEquipmentDisabled)
             {
-                Main.ModLogger.LogError("Swepswep will not load because you disabled an item or equipment. Shame on you.");
+                Main.ModLogger.LogError("Swepswep the Sandy will not load because you disabled an item or equipment. Shame on you.");
                 return;
             }
 
             var master = PrefabAPI.InstantiateClone(Paths.GameObject.ScavLunar1Master, "SwepSwepMaster", true);
             var body = PrefabAPI.InstantiateClone(Paths.GameObject.ScavLunar1Body, "SwepSwepBody", true);
+            var modelLocator = body.GetComponent<ModelLocator>();
+            var modelTransform = modelLocator._modelTransform;
+            var scavMesh = modelTransform.Find("ScavMesh").GetComponent<SkinnedMeshRenderer>();
+
+            var newScavMeshMat = new Material(Paths.Material.matScavLunar);
+            newScavMeshMat.SetColor("_TintColor", new Color32(162, 137, 69, 255));
+            newScavMeshMat.SetColor("_EmColor", new Color32(203, 99, 0, 255));
+            newScavMeshMat.SetFloat("_EmPower", 0.2f);
+            newScavMeshMat.SetFloat("_GreenChannelBias", 0.38f);
+            scavMesh.material = newScavMeshMat;
+
+            var scavEyeMesh = modelTransform.Find("ScavEyeMesh").GetComponent<SkinnedMeshRenderer>();
+
+            var newScavEyeMeshMat = new Material(Paths.Material.matScavLunarEyes);
+            newScavEyeMeshMat.SetColor("_Color", new Color32(187, 5, 0, 255));
+            newScavEyeMeshMat.SetFloat("_EmPower", 0.25f);
+
+            scavEyeMesh.material = newScavEyeMeshMat;
+
+            var scavBackpackMesh = modelTransform.Find("ScavBackpackMesh").GetComponent<SkinnedMeshRenderer>();
+
+            var newScavBackpackMeshMat = new Material(Paths.Material.matScavLunarBackpack);
+            newScavBackpackMeshMat.SetColor("_TintColor", Color.white);
+            newScavBackpackMeshMat.SetColor("_EmColor", new Color32(123, 62, 0, 255));
+            newScavBackpackMeshMat.SetFloat("_EmPower", 0.1f);
+
+            scavBackpackMesh.material = newScavBackpackMeshMat;
+
+            var rootBase = modelTransform.Find("ScavArmature/ROOT/base");
+            var scavWeaponMesh = rootBase.Find("WeaponParent/ScavWeaponMesh").GetComponent<MeshRenderer>();
+
+            var newScavWeaponMeshMat = new Material(Paths.Material.matTrimSheetConstructionBlueEmission);
+            newScavWeaponMeshMat.SetColor("_TintColor", Color.white);
+            newScavWeaponMeshMat.SetColor("_EmColor", new Color32(213, 107, 0, 255));
+            newScavWeaponMeshMat.SetFloat("_EmPower", 0.1f);
+
+            scavWeaponMesh.material = newScavWeaponMeshMat;
 
             var characterMaster = master.GetComponent<CharacterMaster>();
             characterMaster.bodyPrefab = body;